score entierement fonctionnel (sur 1 niveau pour l'instant). Voir si on peut pas faire un arbre des scores

+++ AstronautComplex/ExerciceScore.cs
@@ -12,7 +12,7 @@ namespace AstronautComplex
 
         public override string ToString()
         {
-            return (GoodAndswers / TotalAnswers) * 100 + " %";
+            return ((double)GoodAndswers / TotalAnswers) * 100 + " %";
         }
     }
 }
